# docs/swagger.yaml
definitions:
  dto.MembershipEventDTO:
    properties:
      amount:
        type: string
      end_duration:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      status:
        type: integer
      transaction_hash:
        type: string
      user_address:
        type: string
    type: object
  dto.TransferTokenPayloadDTO:
    properties:
      recipient_address:
        type: string
      token_amount:
        type: string
      tx_type:
        type: string
    type: object
  util.GeneralError:
    properties:
      code:
        description: HTTP error code, or custom error code
        type: integer
      errors:
        description: List of error send server 2 server
        items:
          type: string
        type: array
      string:
        description: Friendly error message
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/membership/events:
    get:
      consumes:
      - application/json
      description: Get a list of membership events based on provided order IDs
      parameters:
      - description: Comma-separated list of Order IDs
        in: query
        name: orderIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MembershipEventDTO'
            type: array
        "400":
          description: Invalid Order IDs
          schema:
            $ref: '#/definitions/util.GeneralError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.GeneralError'
      summary: Get membership events by order IDs
      tags:
      - membership
  /api/v1/transfer:
    post:
      consumes:
      - application/json
      description: Transfer
      parameters:
      - description: Request transfer tokens, required
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.TransferTokenPayloadDTO'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 'When success, return {"success": true}'
          schema:
            items:
              $ref: '#/definitions/dto.TransferTokenPayloadDTO'
            type: array
        "400":
          description: Invalid payload
          schema:
            $ref: '#/definitions/util.GeneralError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.GeneralError'
      summary: Transfer
      tags:
      - transfer
swagger: "2.0"